Refactor parallel test
